+++ packages/themes/src/bitcasinoDark/bitcasinoDark.ts
@@ -8,45 +8,36 @@ const boxShadow =
   '0px 2.4px 7.2px rgba(0, 0, 0, 0.18), 0px 12.8px 28.8px rgba(0, 0, 0, 0.22)';
 
 const color = {
-  piccolo: {
-    120: '#02C556',
-    100: '#0CD463',
-    80: '#19E170',
-  },
-  hit: {
-    120: '#11161D',
-    100: '#1A212A',
-    80: '#242D38',
-  },
-  beerus: {
-    100: '#31373F',
-  },
   goku: {
-    100: '#232A33',
-    80: '#353C44',
-    40: '#4C5159',
-    10: '#62676D',
+    100: '#24033F',
+    80: '#320854',
+    60: '#3F1065',
+    40: '#4C1A75',
+    10: '#65348D',
   },
   gohan: {
-    100: '#1A212A',
-    80: '#31373F',
-    40: '#5F646A',
-    10: '#8C9095',
+    100: '#3E0D68',
+    80: '#4B1679',
+    60: '#582187',
+    40: '#642E94',
+    20: '#713C9F',
   },
-  goten: {
-    100: '#FFFFFF',
+  beerus: {
+    100: '#521B81',
   },
   bulma: {
     100: '#FFFFFF',
   },
   trunks: {
-    100: '#8697A2',
+    100: '#A18EB4',
+    60: '#8D71A8',
+    20: '#795699',
   },
 };
 
-const sportsbetDark: Theme = {
+const bitcasinoDark: Theme = {
   ...sharedTokens,
-  brand: 'Sportsbet.io',
+  brand: 'Bitcasino.io',
   fontFace: avertaStd.fontFace,
   fontFamily: avertaStd.fontStack,
   fontSize: {
@@ -58,9 +49,22 @@ const sportsbetDark: Theme = {
     text: color.bulma[100],
     background: color.goku[100],
     ...supportColors,
+    piccolo: {
+      120: '#F54100',
+      100: '#FF4F0F',
+      80: '#FA642E',
+    },
+    hit: {
+      120: '#0679E5',
+      100: '#1286F3',
+      80: '#2F92EE',
+    },
+    goten: {
+      100: '#FFFFFF',
+    },
     ...color,
   },
   colorScheme: 'dark',
 };
 
-export default sportsbetDark;
+export default bitcasinoDark;