+++ packages/themes/src/supportColors/supportColors.ts
@@ -1,0 +1,45 @@
+import { ColorShared } from '../types/color';
+
+const supportColors: ColorShared = {
+  // Warning colors
+  krillin: {
+    100: '#FFB319',
+    10: '#FFE4AD',
+  },
+  // Error colors
+  chiChi: {
+    100: '#FF4E64',
+    10: '#FFA7B2',
+  },
+  // Success colors
+  roshi: {
+    100: '#0CD664',
+    10: '#B6F3D0',
+  },
+  dodoria: {
+    100: '#D33030',
+    10: '#E99797',
+  },
+  cell: {
+    100: '#0AB758',
+    10: '#84DBAB',
+  },
+  raditz: {
+    100: '#B3804A',
+    10: '#D9BFA4',
+  },
+  whis: {
+    100: '#0F75BD',
+    10: '#87BADE',
+  },
+  frieza: {
+    100: '#5C33CF',
+    10: '#B6A3E9',
+  },
+  nappa: {
+    100: '#725550',
+    10: '#B8AAA7',
+  },
+};
+
+export default supportColors;